21/2

+++ src/main/java/com/example/demo/model/Grammar.java
@@ -8,7 +8,7 @@ public class Grammar {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(name = "grammar_name", nullable = false, length = 255)
     private String grammarName;
@@ -31,11 +31,11 @@ public class Grammar {
         this.contentMarkDown = contentMarkDown;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
